user-management-api/requirements.txt
botocore
fastapi
boto3
httpx
pydantic[email]
python-jose[cryptography]
python-multipart
mangum
uvicorn[standard]
pyjwt[crypto]
requests